% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundCollection-package.R
\docType{package}
\name{compoundCollection-package}
\alias{compoundCollection-package}
\alias{compoundCollection}
\title{Annotation and Structure Collection of Small Molecules}
\description{
This package contains annotation and structures of drugs/small
molecules in DrugAge, DrugBank, CMAP02 and LINCS databases, it also supports
adding custom compound collections.
}
\author{
\itemize{
  \item Yuzhu Duan (yduan004@ucr.edu)
  \item Thomas Girke (thomas.girke@ucr.edu)
}
}
