% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customCompoundDB-package.R
\docType{package}
\name{customCompoundDB-package}
\alias{customCompoundDB-package}
\alias{customCompoundDB}
\title{Query Interface for compoundCollectionData package}
\description{
This software package is the query interface for the corresponding
\code{\link[compoundCollectionData]{compoundCollectionData}} package.
It contains utilities to query the compound annotations in DrugAge, DrugBank,
CMAP02 and LINCS databases by providing ChEMBL ids of the query compounds.
It also supports adding custom compound annotations to the annotation SQLite
database in the \code{compoundCollectionData} package.
}
\seealso{
\code{\link{customAnnot}}, \code{\link{queryAnnotDB}}
}
\author{
\itemize{
  \item Yuzhu Duan (yduan004@ucr.edu)
  \item Thomas Girke (thomas.girke@ucr.edu)
}
}
