% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fcts.R
\name{queryAnnotDB}
\alias{queryAnnotDB}
\title{Annotation Queries with ChEMBL IDs}
\usage{
queryAnnotDB(chembl_id, annot = c("DrugAge", "DrugBank", "CMAP02", "LINCS"))
}
\arguments{
\item{chembl_id}{character vector of ChEMBL IDs.}

\item{annot}{character vector of annotation resources, such as
\code{DrugAge}, \code{DrugBank}, \code{CMAP02}, \code{LINCS} or names of
the annotation tables added by users}
}
\value{
data.frame of annotation result
}
\description{
This function can be used to query compound annotations from the default
resources as well as the custom resources stored in the SQLite annotation
database. The default annotation resources are DrugAge, DrugBank, CMAP02 and
LINCS. The detailed description of the SQLite database is available at the
corresponding \code{\link[compoundCollectionData]{compoundCollectionData}}
package. The custom compound annotations could be added/deleted by the
\code{\link{customAnnot}} utilities.
The input of this query function is a set of ChEMBL IDs, it returns a
data.frame storing annotations of the input compounds from selected
annotation resources defined by the \code{annot} argument.
}
\examples{
query_id <- c("CHEMBL1000309", "CHEMBL100014", "CHEMBL100109",
               "CHEMBL100", "CHEMBL1000", "CHEMBL10")
annot_res <- queryAnnotDB(query_id, annot=c("DrugAge", "LINCS"))

# Add a custom compound annotation table
chembl_id <- c("CHEMBL1000309", "CHEMBL100014", "CHEMBL10",
               "CHEMBL100", "CHEMBL1000", NA)
annot_tb <- data.frame(compound_name=paste0("name", 1:6),
        chembl_id=chembl_id,
        feature1=paste0("f", 1:6),
        feature2=rnorm(6))
addCustomAnnot(annot_tb, annot_name="mycustom")

# query custom annotation
annot_res2 <- queryAnnotDB(query_id, annot=c("LINCS", "mycustom"))
}
