% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fcts.R
\name{queryCollection}
\alias{queryCollection}
\title{Annotation Queries with ChEMBL IDs
Query drug annotations stored in the SQLite database (the defaults are DrugAge,
DrugBank, CMAP02, LINCS) with ChEMBL IDs.}
\usage{
queryCollection(chembl_id, annot = c("DrugAge", "DrugBank", "CMAP02", "LINCS"))
}
\arguments{
\item{chembl_id}{character vector of ChEMBL IDs.}

\item{annot}{character vector of annotation resource, such as DrugAge,
DrugBank, CMAP02 or LINCS}
}
\value{
data.frame of annotation result
}
\description{
Annotation Queries with ChEMBL IDs
Query drug annotations stored in the SQLite database (the defaults are DrugAge,
DrugBank, CMAP02, LINCS) with ChEMBL IDs.
}
\examples{
chembl_id <- c("CHEMBL1064", "CHEMBL10", "CHEMBL113", "CHEMBL1004", "CHEMBL31574")
anno_res <- queryCollection(chembl_id, annot=c("DrugAge", "LINCS"))

}
