% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customCompoundDB-package.R
\docType{package}
\name{customCompoundDB-package}
\alias{customCompoundDB-package}
\alias{customCompoundDB}
\title{Customize and Query Compound Annotation Database}
\description{
This software package is the customization and query interface for the
annotation SQLite database from the corresponding
\code{\link[compoundCollectionData]{compoundCollectionData}} package.
It provides utilities to query the compound annotations from DrugAge, DrugBank,
CMAP02 and LINCS resources by providing ChEMBL ids of the query compounds.
It also supports adding customized compound annotations to the annotation
SQLite database.
}
\seealso{
\code{\link{customAnnot}}, \code{\link{queryAnnotDB}}
}
\author{
\itemize{
  \item Yuzhu Duan (yduan004@ucr.edu)
  \item Thomas Girke (thomas.girke@ucr.edu)
}
}
